Migration 'E_ChangedWeightsDatatype'

+++ FitLibrary.DataAccess.Common/Models/ExerciseDb.cs
@@ -7,7 +7,7 @@
         public string ExerciseName { get; set; }
         public int? Sets { get; set; }
         public int? Repetitions { get; set; }
-        public double? Weight { get; set; }
+        public decimal? Weight { get; set; }
         public string Time { get; set; } = null!;
         public string RestPeriod { get; set; } = null!;
         public int TrainingPlanId { get; set; }
